refactor: split api.go into api_resource.go chunks
